+++ src/data/models/moussaillon-rights.ts
